updated default allocator

+++ kernel/mem/defaultalloc/defaultalloc.hpp
@@ -5,21 +5,21 @@ namespace Heap
 {
     struct HeapSegHdr
     {
-        size_t Length;
+        uint64_t Length;
         HeapSegHdr *Next;
         HeapSegHdr *Last;
         bool IsFree;
         void CombineForward();
         void CombineBackward();
-        HeapSegHdr *Split(size_t SplitLength);
+        HeapSegHdr *Split(uint64_t SplitLength);
     } __attribute__((aligned(16)));
 }
 
 #define defPREFIX(func) defk##func
 
-void InitHeap(void *HeapAddress, size_t PageCount);
+void InitHeap(void *HeapAddress, uint64_t PageCount);
 
 extern void defPREFIX(free)(void *Address);
-extern void *defPREFIX(malloc)(size_t Size);
-extern void *defPREFIX(calloc)(size_t n, size_t Size);
-extern void *defPREFIX(realloc)(void *Address, size_t Size);
+extern void *defPREFIX(malloc)(uint64_t Size);
+extern void *defPREFIX(calloc)(uint64_t n, uint64_t Size);
+extern void *defPREFIX(realloc)(void *Address, uint64_t Size);
